Locking down mutations requiring authorization

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
+import getUserId from "../utils/getUserId";
 
 const Mutation = {
   async createUser(parent, args, { prisma }, info) {
@@ -43,35 +44,54 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, "thisisasecret"),
     };
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    return prisma.mutation.deleteUser({ where: { id: args.id } }, info);
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    return prisma.mutation.deleteUser({ where: { id: userId } }, info);
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id,
+          id: userId,
         },
         data: args.data,
       },
       info
     );
   },
-  createPost(parent, args, { prisma }, info) {
-    return prisma.mutation.createPost({
-      data: {
-        title: args.data.title,
-        body: args.data.body,
-        published: args.data.published,
-        author: {
-          connect: {
-            id: args.data.author,
+  createPost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    return prisma.mutation.createPost(
+      {
+        data: {
+          title: args.data.title,
+          body: args.data.body,
+          published: args.data.published,
+          author: {
+            connect: {
+              id: userId,
+            },
           },
         },
       },
-    });
+      info
+    );
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+
+    if (!postExists) {
+      throw new Error("Cannot edit post.");
+    }
+
     return prisma.mutation.updatePost(
       {
         where: {
@@ -82,7 +102,20 @@ const Mutation = {
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+
+    if (!postExists) {
+      throw new Error("Unable to delete post.");
+    }
+
     return prisma.mutation.deletePost(
       {
         where: {
@@ -92,14 +125,16 @@ const Mutation = {
       info
     );
   },
-  createComment(parent, args, { prisma }, info) {
+  createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author,
+              id: userId,
             },
           },
           post: {
@@ -112,7 +147,20 @@ const Mutation = {
       info
     );
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+
+    if (!commentExists) {
+      throw new Error("Cannot update comment.");
+    }
+
     return prisma.mutation.updateComment(
       {
         where: {
@@ -123,7 +171,20 @@ const Mutation = {
       info
     );
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId,
+      },
+    });
+
+    if (!commentExists) {
+      throw new Error("Cannot delete comment.");
+    }
+
     return prisma.mutation.deleteComment(
       {
         where: {
